Fix possible NPE and aloow to change Beaker URL
When beaker builder is installed for the first time, beaker server URL is not setup and therefore beaker client is not setup (is null).
Later on, when URL is setup in main config, beaker client is not recreated for that to get rid of null, Jenkins has to be restarted.
Also when one want to change Beaker URL, it's not changed and Jenkins needs to be restarted.